+++ R/make_gt_labels.R
@@ -9,6 +9,9 @@
 #' @param format_cols_number Character vector of columns to format as numeric (e.g., mean).
 #' @param format_cols_integer Character vector of columns to format as integer (e.g., minimum, maximum).
 #' @param format_cols_percent Character vector of columns to format as percentage (e.g., percent).
+#' @param decimals_number Numeric vector of number of decimals to use for numeric values.
+#' @param decimals_integer Numeric vector of number of decimals to use for integer values.
+#' @param decimals_percent Numeric vector of number of decimals to use for percent values.
 #' @param summarize_cols_total Character vector of columns to summarize in total rows.
 #' @param summarize_cols_mean Character vector of columns to summarize in mean rows.
 #'
@@ -22,6 +25,9 @@ make_gt_labels <- function(
     format_cols_number = NULL,
     format_cols_integer = NULL,
     format_cols_percent = NULL,
+    decimals_number = NULL,
+    decimals_integer = NULL,
+    decimals_percent = NULL,
     summarize_cols_total = NULL,
     summarize_cols_mean = NULL
 ) {
@@ -34,6 +40,11 @@ make_gt_labels <- function(
   format_cols_number <- unique(c("mean", format_cols_number))
   format_cols_integer <- unique(c("total", "count", "minimum", "maximum", format_cols_integer))
   format_cols_percent <- unique(c("percent", format_cols_percent))
+
+  decimals_number <- 0
+  decimals_integer <- 1
+  decimals_percent <- 0
+
   ignore_cols <- unique(c("empty", group_name, ignore_cols))
 
   summarize_cols_total <- unique(c(format_cols_integer, summarize_cols_total))
@@ -51,6 +62,9 @@ make_gt_labels <- function(
       format_cols_number = format_cols_number,
       format_cols_integer = format_cols_integer,
       format_cols_percent = format_cols_percent,
+      decimals_number = decimals_number,
+      decimals_integer = decimals_integer,
+      decimals_percent = decimals_percent,
       summarize_cols_total = summarize_cols_total,
       summarize_cols_mean = summarize_cols_mean,
       label_group = label_group,