+++ NEWS.md
@@ -1,3 +1,9 @@
+# utils.gt 0.1.2
+
+* Enable setting number of decimals when formatting a number with make_gt_labels.
+* Apply a lighter style to zero values with style_gt_zeros.
+* Add mesocarnivore to omnivore style in style_gt_diet.
+
 # utils.gt 0.1.1
 
 * Enable revision of gt_labels with named list.